Move selector functions to proper place
