+++ src/reducers/observations.js
@@ -94,42 +94,3 @@ module.exports.default = function (state = initialState, action) {
   /* default */
   } else return state
 }
-
-/*
- * Get a FeatureCollection of all active Observations
- */
-module.exports.getActiveFeatures = function (state) {
-  const all = state.get('_map')
-  const activeFeatures = []
-  state.get('active').forEach(id => activeFeatures.push(all[id]))
-  return {
-    type: 'FeatureCollection',
-    features: activeFeatures
-  }
-}
-
-/*
- * Get a flattened, non-immutable map of Observation properties
- * eg.
- * {
- *  propertyName: {
- *    response1: response1_count,
- *    response2: response2_count
- *  }
- * }
- */
-module.exports.getFlattenedProperties = function (state) {
-  const _map = state.get('_map')
-  const flattened = {}
-  state.get('all').forEach(id => {
-    let { properties } = _map[id]
-    for (let propertyName in properties) {
-      let response = properties[propertyName]
-      flattened[propertyName] = flattened[propertyName] || {}
-      let count = flattened[propertyName][response]
-        ? flattened[propertyName][response] + 1 : 1
-      flattened[propertyName][response] = count
-    }
-  })
-  return flattened
-}
